Reorganise profiles into dirs for each provider
As more profiles are added it will be hard to keep track of all the
profiles if they are stored in one directory. This commit creates a
directory per provider and moves the profile files to the appropriate
folder.

A second change is that the version of the OS created by the profile is
also included in the filename.

The only exception to this is the Azure Ubuntu profile which only
specifies the "UbuntuServer" VM image (without a version). At the moment
this happens to be v16.04 but it is my understanding that this could
change in the future. A solution to this would be to specify the exact
Ubuntu version in the profile.

Closes #453
