+++ examples/digitalocean/basiccluster.go
@@ -18,12 +18,12 @@ import (
 	"github.com/kris-nova/kubicorn/cutil"
 	"github.com/kris-nova/kubicorn/cutil/initapi"
 	"github.com/kris-nova/kubicorn/cutil/logger"
-	"github.com/kris-nova/kubicorn/profiles"
+	"github.com/kris-nova/kubicorn/profiles/digitalocean"
 )
 
 func main() {
 	logger.Level = 4
-	cluster := profiles.NewUbuntuDigitalOceanCluster("myCluster")
+	cluster := digitalocean.NewUbuntuCluster("myCluster")
 	cluster, err := initapi.InitCluster(cluster)
 	if err != nil {
 		panic(err.Error())
